HSEARCH-4229 Use Long for OutboxEvent identifiers

+++ integrationtest/mapper/orm/src/test/java/org/hibernate/search/integrationtest/mapper/orm/automaticindexing/strategy/outbox/OutboxPollingAutomaticIndexingStrategyEdgeIT.java
@@ -232,7 +232,7 @@ public class OutboxPollingAutomaticIndexingStrategyEdgeIT {
 		} );
 
 		// Remember the events at this point
-		List<Integer> eventIdsUpToDelete = new ArrayList<>();
+		List<Long> eventIdsUpToDelete = new ArrayList<>();
 		withinTransaction( sessionFactory, session -> {
 			eventIdsUpToDelete.addAll( outboxEventFinder.findOutboxEventIdsNoFilter( session ) );
 		} );
